Refactor configuration setup into a dedicated module
